update(flaskr): Add conftest.py for fixtures.
Abstracts out the fixtures into a separate conftest.py file to enable fixture reuse.
